allow floats as input

+++ cmd/gpud/main.go
@@ -6,8 +6,8 @@ import (
 	"github.com/an0nymoos3/gpud/internals/gpupricing"
 )
 
-func getInput() int {
-	var usd int
+func getInput() float32 {
+	var usd float32
 
 	fmt.Print("Please enter price in USD: ")
 	fmt.Scan(&usd)
@@ -17,7 +17,7 @@ func getInput() int {
 func main() {
 	usd_price := getInput()
 	gpu_price := gpupricing.GetPrice()
-	gpud := float32(usd_price) / float32(gpu_price)
+	gpud := usd_price / float32(gpu_price)
 
-	fmt.Printf("$%d USD = $%f GPUD \n", usd_price, gpud)
+	fmt.Printf("$%f USD = $%f GPUD \n", usd_price, gpud)
 }
